moved GenRegDescJson
