+++ src/main/scala/util/Annotations.scala
@@ -3,8 +3,11 @@
 package freechips.rocketchip.util
 
 import Chisel._
-import chisel3.experimental.{annotate, ChiselAnnotation, RawModule}
+import chisel3.experimental.{ChiselAnnotation, RawModule, annotate}
 import firrtl.annotations._
+import freechips.rocketchip.regmapper.RegField
+import org.json4s.JsonDSL._
+import org.json4s.jackson.JsonMethods.{pretty, render}
 
 // TODO: replace this with an implicit class when @chisel unprotects dontTouchPorts
 trait DontTouch { self: RawModule =>
@@ -38,4 +41,26 @@ case class RetimeModuleAnnotation(target: ModuleName) extends SingleTargetAnnota
 /** Mix this into a Module class or instance to mark it for register retiming */
 trait ShouldBeRetimed { self: RawModule =>
   chisel3.experimental.annotate(new ChiselAnnotation { def toFirrtl = RetimeModuleAnnotation(self.toNamed) })
+}
+
+
+
+object GenRegDescJson {
+
+  def serialize(base: BigInt, name: String, mapping: RegField.Map*): String = {
+
+
+    val regDescs = mapping.flatMap { case (byte, seq) =>
+      seq.map(_.width).scanLeft(0)(_ + _).zip(seq).map { case (bit, f) =>
+        val anonName = s"unnamedRegField${byte.toHexString}_${bit}"
+        (f.desc.map{ _.name}.getOrElse(anonName)) -> f.toJson(byte, bit)
+      }
+    }
+
+    pretty(render(
+      ("peripheral" -> (
+        ("displayName" -> name) ~
+          ("baseAddress" -> s"0x${base.toInt.toHexString}") ~
+          ("regfields" -> regDescs)))))
+  }
 }
